Created Menu buttons

+++ src/Entities/Entity.java
@@ -23,7 +23,7 @@ public abstract class Entity
         g.drawRect((int) hitbox.x, (int) hitbox.y, (int) hitbox.width, (int) hitbox.height);
     }
 
-    protected void intitHitbox(float x, float y, float width, float height)
+    protected void intitHitbox(float x, float y, int width, int height)
     {
         hitbox = new Rectangle2D.Float(x, y, width, height);
     }
